Move exception classes to domain/exception package
Move PauserException and all child exception classes from
com.scalar.admin.kubernetes to com.scalar.admin.kubernetes.domain.exception
package to establish proper domain layer structure.

This commit only moves the files and updates package declarations and
imports. No functional changes or additions (e.g., Javadoc) are included
to ensure git properly tracks the file history as renames.

Changes:
- Move 6 exception classes using git mv (preserves git history)
- Update package declarations
- Update imports in affected classes (Pauser, TlsPauser, TargetSelector)
- Update test imports (PauserTest, TargetSelectorTest)

Moved files:
- GetTargetAfterPauseFailedException.java
- PauseFailedException.java
- PauserException.java
- StatusCheckFailedException.java
- StatusUnmatchedException.java
- UnpauseFailedException.java
