+++ lib/src/main/java/com/scalar/admin/kubernetes/Pauser.java
@@ -3,6 +3,12 @@ package com.scalar.admin.kubernetes;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.util.concurrent.Uninterruptibles;
 import com.scalar.admin.RequestCoordinator;
+import com.scalar.admin.kubernetes.domain.exception.GetTargetAfterPauseFailedException;
+import com.scalar.admin.kubernetes.domain.exception.PauseFailedException;
+import com.scalar.admin.kubernetes.domain.exception.PauserException;
+import com.scalar.admin.kubernetes.domain.exception.StatusCheckFailedException;
+import com.scalar.admin.kubernetes.domain.exception.StatusUnmatchedException;
+import com.scalar.admin.kubernetes.domain.exception.UnpauseFailedException;
 import io.kubernetes.client.openapi.Configuration;
 import io.kubernetes.client.openapi.apis.AppsV1Api;
 import io.kubernetes.client.openapi.apis.CoreV1Api;
