+++ lib/src/main/java/com/scalar/admin/kubernetes/TargetSelector.java
@@ -1,5 +1,6 @@
 package com.scalar.admin.kubernetes;
 
+import com.scalar.admin.kubernetes.domain.exception.PauserException;
 import io.kubernetes.client.openapi.ApiException;
 import io.kubernetes.client.openapi.apis.AppsV1Api;
 import io.kubernetes.client.openapi.apis.CoreV1Api;
