+++ lib/src/main/java/com/scalar/admin/kubernetes/domain/exception/GetTargetAfterPauseFailedException.java
@@ -1,4 +1,4 @@
-package com.scalar.admin.kubernetes;
+package com.scalar.admin.kubernetes.domain.exception;
 
 public class GetTargetAfterPauseFailedException extends PauserException {
   public GetTargetAfterPauseFailedException(String message) {
