+++ lib/src/main/java/com/scalar/admin/kubernetes/domain/exception/UnpauseFailedException.java
@@ -1,4 +1,4 @@
-package com.scalar.admin.kubernetes;
+package com.scalar.admin.kubernetes.domain.exception;
 
 public class UnpauseFailedException extends PauserException {
   public UnpauseFailedException(String message) {
